fix warnings found when compiling with C++
git-svn-id: https://latex2rtf.svn.sourceforge.net/svnroot/latex2rtf/trunk@1090 e562ef82-0334-0410-8083-dbe22b1e9a75

+++ funct1.c
@@ -1497,7 +1497,7 @@ void CmdSpace(float kk)
  purpose: inserts a space of width kk*space 
  ******************************************************************************/
 {
-    int size = CurrentFontSize() * kk;
+    int size = (int) (CurrentFontSize() * kk);
 
     if (getTexMode() == MODE_VERTICAL)
         changeTexMode(MODE_HORIZONTAL);
